requirements.txt
dnspython
requests
tabulate
pyfiglet
